provide common gateway client mock impl
